refactor: extract wav-utils crate shared by orchestrator + vad
Both crates carried a copy of `wav_from_pcm_s16le_mono` (44-byte
WAV header for whisper.cpp's multipart upload). The previous "not
worth the workspace churn" comment is no longer true now that we
have a follow-up review item asking for a single source of truth.

Move the canonical implementation into a new tiny `wav-utils`
crate at the repo root. Both consumers depend on it via path so
each keeps its own Cargo.lock and Docker build cache; we don't
introduce a Cargo workspace because the dependency tree of the
two crates is otherwise disjoint and a workspace lock would force
consistency we don't actually want.

Dockerfiles: build context for orchestrator + VAD switches to the
repo root so the path dep resolves at compile time. compose.yaml
plus the three test-stack composes (asr offline/online + orch
integration) updated accordingly.

Addresses review item 5.
