+++ orchestrator/Cargo.toml
@@ -26,6 +26,14 @@ anyhow = "1"
 clap = { version = "4", features = ["derive", "env"] }
 base64 = "0.22"
 reqwest = { version = "0.12", default-features = false, features = ["json", "multipart", "rustls-tls"] }
+wav-utils = { path = "../wav-utils" }
+
+[dev-dependencies]
+# Tests spin up a tokio TCP listener to mock ollama's /api/chat
+# stream end-to-end. The runtime tokio dep doesn't enable
+# `net` / `io-util` directly (axum + reqwest pull them transitively
+# at build time), but the test module needs to name them.
+tokio = { version = "1", features = ["net", "io-util"] }
 
 [profile.release]
 lto = "thin"
