+++ voice-activity-detection/Cargo.toml
@@ -37,6 +37,11 @@ nnnoiseless = { version = "0.5", default-features = false }
 # 48 kHz only and our pipeline is 16 kHz — one FftFixedInOut
 # handles the 16k→48k upsample, a second handles 48k→16k.
 rubato = "0.16"  # 0.x API stable + matches the FftFixedInOut docs we coded against
+# Shared 44-byte WAV header helper. The orchestrator owns the
+# canonical implementation in `../wav-utils/`; both crates depend on
+# it via path so a future change to the WAV envelope (e.g. adopting
+# WAVEFORMATEXTENSIBLE) lands in one place.
+wav-utils = { path = "../wav-utils" }
 
 [profile.release]
 lto = "thin"
